requirements.txt
argcomplete==1.8.1
argh==0.26.2
asgiref==3.2.7
asn1crypto==0.24.0
blessings==1.6
bpython==0.17.1
catfish==1.4.4
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
curtsies==0.2.12
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.0.5
django-cors-headers==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
ewmh==0.1.6
greenlet==0.4.12
httplib2==0.9.2
idna==2.6
jedi==0.16.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
menulibre==2.2.0
mugshot==0.4.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
onboard==1.4.1
parso==0.6.2
pathtools==0.1.2
pexpect==4.2.1
Pillow==7.1.1
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygame==1.9.6
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.7.1
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-debianbts==2.7.2
python-magic==0.4.16
python-xlib==0.26
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
sgt-launcher==0.2.4
simplejson==3.13.2
six==1.14.0
sqlparse==0.3.1
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
vboxapi==1.0
wadllib==1.3.2
watchdog==0.8.3
wcwidth==0.1.7
xcffib==0.5.1
xkit==0.0.0
zope.interface==4.3.2
